Wells-Fargo task 2
Your task is to implement the data model you created in the previous task within the provided Spring application. Follow these steps carefully:

Fork and clone the starter repo, which can be found here. If you are unfamiliar with Git, read through the first two chapters of the Git book, found here. Git is one of a developer’s most frequently used technologies, so it’s worth taking some time to get acquainted with it.
Download and install IntelliJ if you haven’t already. IntelliJ is one of the best Java IDEs on the market and can be found here.
Open up the project in IntelliJ, then take some time to explore the codebase. Get a feel for what each class is doing. It’s okay if you don’t understand how part of the code works; there are lots of gotchas in Spring development. The goal is to become familiar with the overall flow of the program.
Time to get to work! Create a class for each entity in your data model (these should be placed in the entities directory). Respect the following when implementing your data model:
Each entity must be annotated with the @Entity type, which can be found in the javax.persistence package.
Each id must be auto-generated.
Each instance variable must contain either a column or a relationship annotation.
Each class must contain a constructor which initializes all instance variables.
Each class must expose a getter and setter for each instance variable (no setter for the id field is required).
Lean on the existing entities (one has been provided for you) for hints on how to accomplish the above.
When you are finished implementing your data model, commit and push your changes to GitHub. Then, submit a link to your project repo in the next step. Good luck!